Don't fail when encountering a warning about transitive dependencies
The current Haskell builder fails when it encounters the a warning
from Setup about "indirectly depending on multiple versions of the
same pacakge". There is a call to grep which exits if the
warning string is in the output of Setup.

While indirectly depending on different versions of the
same package *can* be a problem, it can also be the desired
behavior. I ran into this limitation when I was intentionally building
a project that indirect dependencies on megaparsec 6 and megaparsec 7
in a way that did not interfere—the different indirect dependency not
only worked but was also the only way I could get the project to
build.

The cases where this *is* a fundamental problem will be caught at
compile time further into the build process with a reasonably readable
error message, so it should be fine to let the warning remain a
warning rather than exiting the build when we encounter it.