pythonPackages.sip: rename the module from PyQt5.sip back to sip
PyQt5 5.11 wants a copy of sip in PyQt5.sip to ensure that sip-generated PyQt5
modules have the correct sip runtime version. This issue is not relevant to
Nixpkgs, therefore PyQt5 may be reverted to use the common sip runtime.

This is an alternative to #52613 that does not break packages that need to import sip.

Fixes the build of tortoisehg.