systemd: Make build options configurable
The binary closure for systemd is quite large in the default
configuration. This patch provides a simple way to turn off build
options that aren't needed by some users, removing the corresponding
dependencies. It also makes it easy to turn on additional features that
NixOS doesn't enable by default.

Since doCheck was disabled, this commit also disables building tests by
default. If the caller requests building tests, then doCheck will be
turned on as well. The tests are probably still just as broken as they
were before, but if someone wants to try them, they can build with
features.tests=true.

The build log from this version is identical to the build log from the
previous version, with the following exceptions:

- Store paths vary, of course.
- Configure messages like "Dependency libselinux found: NO" are gone;
  the corresponding checks are explicitly disabled now.
- Tests aren't built since we don't run them.
- The build phase compiles files in different orders due to parallel
  build.
- The install and fixup phases touch files in different orders, but they
  touch all the same files.

Also, the same set of paths appear in each output before and after this
patch. Their contents differ due to varying store paths so I haven't
checked if there are more subtle differences, but based on the above
results I suspect they're equivalent.