libqrencode: alias to `pkgs.qrencode`.
The packages were almost equal, used the same major version and referred
to the same source without specifying custom build arguments in one of
each.

I decided to keep `qrencode` by default as the `lib` prefix may be
confusing as it wrongly indicates that only libraries are provided by
the package, but `bin/qrencode` is packaged in both of them.

Fixes #53133