lib/types: make _module.args a types.args
Currently, `_module.args` is merged according to the default rules,
which produces interesting results if we try to assign an attrset with
`_module.args.name` to an option with type `attrsOf (submodule foo)`
(i.e., one that already has `name` defined).

This adds a new type `args` that will just merge the topmost definition.