grml-zsh-config: remove unecessary dependencies
The buildInputs that used to be here appear to have been put here in the
hope that they would be installed and made available automatically as a
result of installing this package, but that isn't how buildInputs work.

They could be provided by using substituteInPlace to point the package
towards the Nix versions of these packages, but since this is a shell
configuration, I think it makes more sense (and is easier) to use
whatever versions of those commands are available at runtime.