Revert "Revert "linux-hardened: Disable GCC_PLUGIN_RANDSTRUCT""
This reverts commit c68e8b05f005381ac1caf51bb28282b70242b77b.

RANDSTRUCT currently fails to work with out-of-tree modules, as
evinced by
https://github.com/NixOS/nixpkgs/commit/c68e8b05f005381ac1caf51bb28282b70242b77b#commitcomment-31850284
and https://github.com/NixOS/nixpkgs/issues/53522.

Specifically, loading out-of-tree modules results in modsym version
mismatches, as in
   spl: version magic '4.20.0 SMP mod_unload modversions RANDSTRUCT_PLUGIN
from the issue above.

A working hypothesis is that the randstruct seed is not carried over when
building out-of-tree modules but more investigation is needed here.

Closes https://github.com/NixOS/nixpkgs/issues/53522