bear: fix wrapper detection patch by checking result of find_executable
With the wrapper detection patch, if a build invokes an executable that cannot
be found in PATH by `find_executable`, bear will fail with an `AttributeError`
in `os.path.realpath`.

This can happen if the build invokes some project-local tool or command, like
`./build-something`.

Instead of calling using the result of `find_executable` directly, first check
whether the executable was found and fall back to original Bear behavior if
it was not.