ocamlPackages.zarith: use buildOcaml instead of mkDerivation
This has two main advantages:

  - By setting hasSharedObjects = true, buildOcaml will automatically
include a setup-hook.sh that sets CAML_LD_LIBRARY_PATH in dependent
expressions. This is needed to pick up dllzarith.so properly which is
shipped as part of the library.

  - We can kill the ugly assert in the expression and instead change it
to use minimumSupportedOcamlVersion.

(Note: this was reverted in b44d5136e8a8d2d407, but the change is
exactly equivalent -- I wasn't sure what impact zarith might actually
have without checking OfBorg, which I wanted to do first.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>