smartgithg: 18_1_5 -> 18.2.4
This PR is committed using the new version.
The never version requires JRE 10, actually the changelog mentions
downgrading from JDK11 to 10 because of a bug related to https
but this seams to be fixed by now
https://bugs.openjdk.java.net/browse/JDK-8210005

Also I did not find a `jre` output in jdk11, so the closure size
might increase due to it.