overlays: (partially) recursive merge
One of the problems newcomers face when working with overlays is
non-expected behavior of overriding packages in package sets. Let's take,
for example, Python. Here's an overlay:

```
self: super: {
    pythonPackages.my-new-package = super.callPackages ./xxx { };
}
```

Big supprise is that `pythonPackages` set is now empty!

```
$ nix-build -E '(with import ./. { overlays = [
    (self: super: {
        pythonPackages.my-new-package = self.pythonPackages.pip;
    })
]; }; pythonPackages.ipython)'

error: attribute 'ipython' missing, at (string):5:7
```

It is not empty, but contains only one package. That's because attribute
sets are not merged recursively (like NixOS modules do), but are replaced
on each `extends`. Recursive merge is tricky, because it has to force
evaluation in lot of places and breaks the knot.

But we can do poor-man recursive merge "whitelist". Say, when overlay
contains attribute `_merge_pythonPackages = true;`, then we can merge it
recursively:

```
$ nix-build -E '(with import ./. { overlays = [
    (self: super: {
        _merge_pythonPackages = true;
        pythonPackages.my-new-package = self.pythonPackages.pip;
    })
]; }; pythonPackages.ipython)'

/nix/store/k4wf0244qq8xcml85n45sdwalizy0i6k-python2.7-ipython-5.8.0
```

We can also support this attribute in left-hand side overlay (for
example, `all-packages.nix`):

```
  ...
  _merge_pythonPackages = true;
  pythonPackages = python.pkgs;
  _merge_python2Packages = true;
  python2Packages = python2.pkgs;
  _merge_python3Packages = true;
  python3Packages = python3.pkgs;
  ...
```

So we get merge behavior by default for `pythonPackages`. We can also
disable merging in our overlay if desired:
```
$ nix-build -E '(with import ./. { overlays = [
    (self: super: {
        _merge_pythonPackages = false;
        pythonPackages = self.python3Packages;
    })
]; }; pythonPackages.ipython)'

/nix/store/0yx0q6v6nisx4fq1vj8x2vc51n7aj5pz-python3.7-ipython-7.1.1
```