Loads qt plugin paths as registered...
This adds the list of Qt plugin paths into builds using `qtbase`.

The plugins list will be added to the Qt plugin path through the
additional patch.

This ensures that no special environment is required to get the desired
plugin path for Qt software.

This should fix all issues where Qt is unable to load its platform
plugin.

This may fix the issue where Qt tries to load a mismatched version of a
plugin.