unison: fix on darwin
Unison is currently broken by default under Darwin due to a build
failure in libglade, which is a transient dependency for unison's gtk2
GUI.

Unison technically has a native macOS UI, so it should in principle be
possible to package it here and enable it instead of the gtk2 UI by
default in darwin, but I struggled to get there. Short of that, a
working unison install with a CLI but no GUI is preferable to no unison
at all on darwin.

The 'UISTYLE=text' option tells the unison makefile to build without any
GUI support. This makes the package flag 'enableX11 = false' effectively
disable any other unison GUI too, as a simple way of supporting the
darwin case above.