[ghc] If we enable integer-gmp, pass information unconditionally
The check for `targetPlatform == hostPlatform` is simply wrong.  If
we build a cross compiler, and *do* enable integer-gmp, we must
pass the `--with-gmp-*` values, otherwise the integer-gmp libraries
package file (the one ghc-pkg / ghc / cabal consult) will miss the
proper paths to the gmp library!

This time around for
- ghc 8.2.2
- ghc 8.4.3
- ghc 8.6.1
- ghc 8.6.2
- ghc 8.6.3

see 66e2367b46f for ghc 8.4.4