qt56.qtvirtualkeyboard: init at 5.6.3
This adds the "missing" qtvirtualkeyboard module of qt56. I just add
this so I can apply (& test) the patches for a CVE in the next commit.
This might seem strange but in case anyone decided to add / use this in
the future we are on the safe(r) side.

(cherry picked from commit 295a210a23b23e165dc2a6f1f845f550ce3a2c74)