overcommit: let installed hooks find the library
Patches overcommit to use the bundlerApp's GEM_HOME and GEM_PATH. This
means nothing has to be installed globally to use overcommit. A side
effect of this is that, if overcommit is installed globally, hooks will
continue to use the version of overcommit that generated them, and will
have to be regenerated to use a different version of overcommit. This
seems reasonable to me.