neovim: add standalone module to configure it
This PR adds a module to typecheck parameters passed to build the wrapper.
It allows to generate

It adds the possibility to configure the haskell provider.
It adds a knob to disable generation of the neovim remote plugin manisfest.

Using `-u` to load the generated config breaks vim expectations about setting $MYVIMRC,
detecting and autoloading .nvimrc / .vimrc.
This PR changes the behavior (not backwards compatible) to fix this.
The configuration is not fully declarative anymore as it will also attempt to load the user configuration.
see here for more details https://github.com/NixOS/nixpkgs/issues/55376.