system.requiredKernelConfig: packages contraints
This lets packages define a meta.requiredKernelConfig to check for some specific
kernel configurations.

Right now this does it only for the booted kernel (hopefully the required config
could be checked against an arbitrary kernel).

The hope is to then add a knob that builds a configuration (thanks to kernel
structured config) satisfying the constraints.