sage: fix fetchSageDiff (#55783)
For some reason I changed it to use `cgit`s `rawdiff` instead of `patch`
in the update to sage 8.6. Probably commited that by accident, at least
I can't remember the reason. Also changed the excludes filter, the
leading slash prevented it from working.

As a result, the cypari2 patch changed. Only didn't notice because it
was cached.

Fixes #55780