openafs_1_8: Include support for Linux 4.20
OpenAFS version 1.8.2 does not have support for Linux 4.20,
meaning that linuxPackages_latest.openafs_1_8 would fail
to build.

This patch adds patches taken from the OpenAFS git to
remove the references to deprecated functions. This has been
tested on Linux 4.20 and Linux 4.14

This patch must be removed when OpenAFS 1.8.3 is released
and improved when Linux 5.0 is added to Nix