applications/profanity: Fix build
Profanity hasn't been able to build since [1]. This patch addresses the
issue by introducing `autoconf-archive` and removing NIX_CFLAGS_COMPILE
as well as LDFLAGS as a consequence since `autoconf-archive` is able to
handle the AX_PYTHON extension used by Profanity's upstream autoconf
file.

Thanks to @aszlig for helping with this.

[1]: https://github.com/NixOS/nixpkgs/commit/efbe87f3ef769aac5e95512609b4759a43109307