linux: Enable support for Self-Encrypting Drives
To quote block/Kconfig:

> Builds Logic for interfacing with Opal enabled controllers.
> Enabling this option enables users to setup/unlock/lock
> Locking ranges for SED devices using the Opal protocol.

Without `BLK_SED_OPAL`, it is impossible to resume from sleep when using
a locked self-encrypting drive.

This configuration option appeared in earlier kernels, but only reached
maturity in 4.14 according to discussion at:

- https://github.com/Drive-Trust-Alliance/sedutil/issues/90 and
- https://github.com/Drive-Trust-Alliance/sedutil/pull/190

This kernel option is enabled in the default kernels shipped with
Fedora, Debian, and other mainstream Linux distributions.