nixos/chrony: add flag to skip initial boot time adjustment
In certain scenarios running Chrony, chronyc waitsync will never return
or will erroneously timeout while performing the initial boot
adjustment. In cases like this, and maybe some other strange, advanced
scenarios, it may be useful to not wait on the initial boot time
adjustment.

The most apparent reason for this that I have encountered is that
'chronyc waitsync' apparently does not understand 'local stratum'
directives, which are used to run local NTP daemons that are
disconnected from any servers/reference clocks (e.g. in an offline
network). In this case, 'waitsync' apparently never returns even though
the clock is synchronized vacuously. But even if this was handled
'correctly' -- in these cases, there is no need to wait on initial
synchronization anyway.

The most useful application of this feature currently is for NixOS
tests: this allows us to run NixOS tests that run NTP daemon nodes
(disconnected from any network/physical clock), which other nodes can
synchronize to.

(It's unclear to me if 'waitsync' not immediately returning in these
cases of offline 'local stratum' servers is a bug or intended behavior.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>