nixos/sshd: implement support for `Match` groups
This patch makes it way harder to screw Match groups in an `sshd` config
up. With a broken SSH, reverting with tools like `nixops` is way harder,
in some cases impossible.

The diff seems larger than it actually is, mainly because I moved all
options that can be used "globally" and in a `Match` group into its own
attribute set that will be merged into `options.services.openssh` and
for the `matches` submodule.

The `Match` config will be appended to `/etc/ssh/sshd_config` using
`mkOrder 9999` to ensure that no `extraConfig` will be appended
afterwards and only applies to the last Match group.

It can be used like this now:

``` nix
{
  services.openssh = {
    enable = true;
    permitRootLogin = "no";
    matches = [
      {
        # the given config only applies to the 10.23.42.0/24 subnet
        match."10.23.42.0/24" = "Address";
        config.permitRootLogin = "yes";
      }
    ];
  };
}
```

Finally extended the `openssh` testcase with another machine to demonstrate
the functionality by permitting root logins only from certain IPs.