python.pkgs.pyqt5: 5.11.3 -> 5.13.0
This upgrades pyqt5 and fixes the build. Note that the QtWebEngine
module is split into `pyqtwebengine` by upstream and us.

If your application gives an error that a Qt module is missing, then
that is most likely because it was not included when building pyqt5.
The solution is to simply add it here to the `(native)BuildInputs`.
Don't forget to add an import check. We should include all the modules
(except qtwebkit) that are needed.