gtk-doc: patch the on-line DocBook XSL paths to the local
This is what's done in Guix,
so `gtkdoc-mkhtml` would in no case even try to reach for the Internet.
Also removed `dblatex`,
because it was never enabled and causes circular dependencies if it is.
The `--disable-scrollkeeper` configure flag isn't recognized by now.
And `doCheck` should be left alone, as long as the tests pass.