pureUsePlatformOptimizations: add stdenv adapter
This allows one to force a compiler to use optimizations for a particular architecture (for example "sandybridge" or "skylake-avx512").
The resulting binaries won't work on the older or incomparible architecture (likely will crash with "Invalid instruction").

Contrary to `impureUseNativeOptimizations' of https://github.com/NixOS/nixpkgs/pull/37600,
this adapter does not break purity and allows to prepare a NixOps deployment specifying exact CPU architecture of each node,
so it might become the default one day.