cc-wrapper: give -march=... flag less priority
Do not override -march= with the value from $NIX_CFLAGS_COMPILE
which is set for whole closure or even for the whole closure
(via platform.gcc.arch), while per-file -march= is used by many
packages (xen, openblas, qt512.qtwebengine, ...) to compile
snippets of code optimized for a particual architecture.

It allows to set ```pureUsePlatformOptimizations```
(https://github.com/NixOS/nixpkgs/pull/56649) adapter globally:

```
self: super: {
  stdenv = self.stdenvAdapters.pureUsePlatformOptimizations super.stdenv;
}
```

resulting in Gentoo-style compilation of the whole system for
a particual architecture (with the exception of those snippets
above, which might targer even higher or incompatible
(for example AMD-specific) architecture and require
custom -march= which must not be shadowed by a derivation-wide
or system-wide -march.