openafs: fix OpenAFS 1.6 on Linux 4.20 (#56862)
openafs: fix OpenAFS 1.6 on Linux 4.20

This patch cherry-picks commits from the OpenAFS git repository to
add support for Linux 4.20 to OpenAFS 1.6.

There do not seem to be patches in the Git repository for full Linux 5.0
support, although I have added the ones which exist so far.

This patch has been tested on Linux 4.4, 4.14, 4.20, and is confirmed to
fail on Linux 5.0. It does not change the drv of the openafs package,
only linuxPackages_*.openafs.