openafs: Add support for Linux 5.0 to openafs_1_8 (#56859)
OpenAFS 1.8.2 does has not yet merged support for Linux 4.20 or 5.0
This patch cherry-picks commits from the git repository, and includes
them in the patches/ directory. This also replaces the amalgamated Linux 4.20
patches I added earlier with the original outputs from `git show`.

This has been tested to build on Linux 4.4, 4.14, 4.20, and 5.0.

This does not change the derivations of openafs_1_8, only
linuxPackages_*.openafs_1_8.