lib: introduce `foreach`, `combined` and `combined map` idioms
This PR adds `foreach` and `combined` functions. They are repainted
versions of following standard lib functions:
- `foreach        == flip map`
- `foreach        == flip mapAttrs` when applied to attrset
- `combined       == concatString`
- `combined "sep" == concatStringSep "sep"`
- `combined map   == flip concatMapStrings`
- `combined "sep" map == flip (concatMapStringsSep "sep")`

The idea behind this rename is to encourage flipped usage of functions
in templaters and other script generators.

Flipped version are often easier to read. They are also more consistent with
existing usage of `optional`, `optionalAttrs`, `mkIf` in sense that
*usually* first argument is shorter (syntatically) than second.