bonnie: fix bon_csv2txt utility
The bon_csv2txt utility is a perl script. By default it is looking forbonnie: fix bon_csv2txt utility

The bon_csv2txt utility is a perl script. By default it is looking for
perl in /usr/bin/perl which doesn't work on nixos.

This changes patches the perl interpreter correctly. But this comes as at a cost of
closure size increase from 34'006'536 bytes to  90'289'456 bytes as perl is now a dependency.

An alterantive would be to patch the file to use /usr/bin/env perl and just rely on a perl interpreter
being present in PATH.

perl in /usr/bin/perl which doesn't work on nixos.

This changes patches the perl interpreter correctly. But this comes as at a cost of
closure size increase as perl is now a dependency.

An alterantive would be to patch the file to use /usr/bin/env perl and just rely on a perl interpreter
being present in PATH.